requirements.txt
Flask
Flask-Cors
chromadb
google-generativeai
python-dotenv
sentence-transformers
beautifulsoup4
argparse
textract